[#75357] tlib: Fix `extract16()`

+++ include/bit_helper.h
@@ -45,7 +45,7 @@ static inline uint64_t deposit64(uint64_t dst_val, uint8_t start, uint8_t length
 static inline uint32_t extract16(uint16_t value, uint8_t start, uint8_t length)
 {
     tlib_assert(start + length <= 16);
-    return (value >> start) & (UINT32_MAX >> (16 - length));
+    return (value >> start) & (UINT16_MAX >> (16 - length));
 }
 
 /* Extracts 'length' bits of 'value' at 'start' bit. */
